requirements.txt
python-dotenv==0.19.1
pdfplumber==0.5.28